// resources/js/Models/Paginated.ts
export interface Paginated<T> {
    current_page: number;
    data: T[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: {
        url: string;
        active: boolean;
        label: string;
    }[];
    next_page_url: string | null;
    path: string;
    perPage: number;
    prev_page_url: string | null
    to: number;
    total: number;
}